+++ Eikona/Tessa/ConnectorBundle/AttributeType/TessaType.php
@@ -13,6 +13,7 @@ use Akeneo\Pim\Structure\Component\AttributeType\AbstractAttributeType;
 class TessaType extends AbstractAttributeType
 {
     const ATTRIBUTE_EXPORT_URL = 'tessa_export_url';
+    const ATTRIBUTE_MAX_DISPLAYED_ASSETS = 'max_displayed_assets';
 
     /**
      * {@inheritdoc}