+++ Eikona/Tessa/ConnectorBundle/Resources/public/js/datagrid/cell/eikona-catalog-tessa-cell.js
@@ -15,28 +15,44 @@ define(
     'underscore',
     'eikona/tessa/connector/datagrid/cell/template',
     'oro/datagrid/string-cell',
-    'routing'
+    'routing',
+    'eikona/tessa/fetcher/attribute'
   ],
-  function (_, fieldTemplate, StringCell, Routing) {
+  function (_, fieldTemplate, StringCell, Routing, AttributeFetcher) {
     return StringCell.extend({
       fieldTemplate: _.template(fieldTemplate),
       render () {
-        const value = this.model.get(this.column.get('name'));
-        const tessaAssetIds = value ? value.split(',') : [];
-        let showUrl = '';
+        const attributeCode = this.column.get('name');
 
-        if (tessaAssetIds.length <= 0) {
-          return this;
-        }
+        AttributeFetcher.getAttribute(attributeCode).then((attributeDefinition) => {
+          let maxDisplayedAssets = parseInt(attributeDefinition.max_displayed_assets, 10);
+          maxDisplayedAssets = Number.isNaN(maxDisplayedAssets) ? 1 : maxDisplayedAssets;
 
-        const firstAssetId = tessaAssetIds[0];
-        showUrl = Routing.generate('eikona_tessa_media_preview', {assetId: firstAssetId});
+          const value = this.model.get(attributeCode);
+          const tessaAssetIds = value ? value.split(',') : [];
+          if (tessaAssetIds.length === 0) {
+            return this;
+          }
 
-        this.$el.empty()
-          .html(this.fieldTemplate({
-            url: showUrl,
-            assets: tessaAssetIds
-          }));
+          const displayedAssets = tessaAssetIds
+            .slice(0, maxDisplayedAssets)
+            .map((tessaAssetId) => ({
+              url: Routing.generate('eikona_tessa_media_preview', {assetId: tessaAssetId}),
+              assetId: tessaAssetId
+            }))
+
+          const additionalAssets = tessaAssetIds
+            .slice(maxDisplayedAssets)
+            .map((tessaAssetId) => ({
+              assetId: tessaAssetId
+            }))
+
+          this.$el.empty()
+            .html(this.fieldTemplate({
+              displayedAssets,
+              additionalAssets
+            }));
+        });
 
         return this;
       }