+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.0 (2021-08-11)
+* [feature] Adjust the amount of assets displayed in product grid and reference entity grid
+* [feature] Added asset download links in API
+
 ## 1.0.1 (2021-07-15)
 * [bugfix] fixed error when publishing a product